Add go server, css, and template
Using a go web server, the html template is served. Ron's names
are put in there through the template.

CSS and image assets are served with the golang server with the
file server function.

Names are being stored in a plain text file, but the methods for
adding and retrieving them from the database are in, just not
hooked up to the tmeplate.
a simple