Add npm, node, typescript, and angularjs
The backend is going to change to use node since that is what
angular tutorials use and I wanted to try out doing a node server.

Typescript is the recommended language to use angular with so that
is being installed along with all of the node server stuff.

Angularjs will be added as a node module.